[feat] Create health check
On this commit we create a health check for the editor service. This
health check is useful to know when the editor is ready to receive
requests. The main issue is that the editor minimizes the client assets
on the first request. This is a problem because if a client does the
request before the CSS file is created a file without content will be
minimized. Then, all requests to this file afterwards will require the
same file minimized.
There is not a straight way to disable the minimization of some files on
Etherpad. It's all or nothing!